# fetch: improve performance of urlHasHttpsScheme