--- a/lib/web/fetch/util.js
+++ b/lib/web/fetch/util.js
@@ -1168,13 +1168,21 @@ function urlIsLocal (url) {
 
 /**
  * @param {string|URL} url
+ * @returns {boolean}
  */
 function urlHasHttpsScheme (url) {
-  if (typeof url === 'string') {
-    return url.startsWith('https:')
-  }
-
-  return url.protocol === 'https:'
+  return (
+    (
+      typeof url === 'string' &&
+      url[5] === ':' &&
+      url[0] === 'h' &&
+      url[1] === 't' &&
+      url[2] === 't' &&
+      url[3] === 'p' &&
+      url[4] === 's'
+    ) ||
+    url.protocol === 'https:'
+  )
 }
 
 /**