before:
```sh
benchmark                              time (avg)             (min … max)       p75       p99      p999
------------------------------------------------------------------------- -----------------------------
urlHasHttpsScheme "http:" String    25.31 ns/iter      (23.7 ns … 263 ns)  24.66 ns   41.4 ns  74.99 ns
urlHasHttpsScheme "https:" String   31.63 ns/iter     (29.23 ns … 120 ns)  30.49 ns  60.29 ns  66.36 ns
urlHasHttpsScheme "http:" Object      927 ps/iter       (443 ps … 186 ns)    955 ps   1.16 ns   8.56 ns
urlHasHttpsScheme "https:" Object     929 ps/iter     (409 ps … 44.81 ns)    955 ps    1.4 ns   8.49 ns
```

after:
```sh
cpu: AMD Ryzen 7 4800H with Radeon Graphics
runtime: node v21.7.1 (x64-linux)

benchmark                              time (avg)             (min … max)       p75       p99      p999
------------------------------------------------------------------------- -----------------------------
urlHasHttpsScheme "http:" String     1.03 ns/iter       (443 ps … 282 ns)   1.13 ns   1.43 ns   8.59 ns
urlHasHttpsScheme "https:" String     931 ps/iter       (409 ps … 153 ns)    955 ps    1.4 ns  10.54 ns
urlHasHttpsScheme "http:" Object      931 ps/iter     (409 ps … 94.12 ns)    955 ps   1.36 ns  12.45 ns
urlHasHttpsScheme "https:" Object     927 ps/iter     (409 ps … 45.01 ns)    955 ps   1.36 ns   8.42 ns
```